requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
chromadb
